2021-01-10/src/tiny-blockchain/pyproject.toml
[tool.poetry]
name = "tiny-blockchain"
version = "0.1.0"
description = ""
authors = ["uta8a <noumodulity@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodomex = "3.9.9"
[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
